+++ crates/import/src/openapi.rs
@@ -27,9 +27,8 @@ use slumber_core::{
     },
     http::HttpMethod,
     template::Template,
-    util::NEW_ISSUE_LINK,
 };
-use slumber_util::ResultTraced;
+use slumber_util::{NEW_ISSUE_LINK, ResultTraced};
 use std::{fs::File, iter, path::Path};
 use strum::IntoEnumIterator;
 use tracing::{debug, error, info, warn};