+++ crates/tui/src/state.rs
@@ -269,7 +269,7 @@ impl LoadedState {
             }
             Message::CollectionEdit => {
                 let path = self.collection_file.path().to_owned();
-                let command = util::get_editor_command(&path)?;
+                let command = TuiContext::get().config.editor_command(&path)?;
                 util::yield_terminal(command, &self.messages_tx)?;
             }
             Message::CopyRequestUrl => self.copy_request_url()?,
@@ -287,13 +287,15 @@ impl LoadedState {
                 )?;
             }
             Message::FileEdit { file, on_complete } => {
-                let command = util::get_editor_command(file.path())?;
+                let command =
+                    TuiContext::get().config.editor_command(file.path())?;
                 util::yield_terminal(command, &self.messages_tx)?;
                 on_complete(file);
             }
             Message::FileView { file, mime } => {
-                let command =
-                    util::get_pager_command(file.path(), mime.as_ref())?;
+                let command = TuiContext::get()
+                    .config
+                    .pager_command(file.path(), mime.as_ref())?;
                 util::yield_terminal(command, &self.messages_tx)?;
                 // Dropping the file deletes it, so we can't do it until after
                 // the command is done