+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
   - [See docs for more](https://slumber.lucaspickering.me/book/api/configuration/index.html#follow_redirects)
 - Add optional `target` argument to `slumber show paths` to show just a single path
   - E.g. `slumber show paths config` prints just the config path
+- Add `--edit` flag to `slumber show config` and `slumber show collection`
+  - This will open the global config/collection file in your configured editor, similar to `git config --edit`
 
 ## Fixed